+++ src/plone/app/multilingual/dx/cloner.py
@@ -1,4 +1,6 @@
 # -*- coding: utf-8 -*-
+import pkg_resources
+
 from plone.app.multilingual.dx.interfaces import ILanguageIndependentField
 from plone.app.multilingual.interfaces import ILanguage
 from plone.app.multilingual.interfaces import ILanguageIndependentFieldsManager
@@ -8,7 +10,11 @@ from plone.dexterity.utils import iterSchemata
 from z3c.relationfield import RelationValue
 from z3c.relationfield.interfaces import IRelationList
 from z3c.relationfield.interfaces import IRelationValue
-from zope.app.intid.interfaces import IIntIds
+try:
+    pkg_resources.get_distribution('zope.initd')
+    from zope.intid.interfaces import IIntIds
+except pkg_resources.DistributionNotFound:
+    from zope.app.intid.interfaces import IIntIds
 from zope.component import getUtility
 from zope.component import queryAdapter
 from zope.interface import implementer